Handle new fastboot slot argument
Old versions of fastboot passed the whole slot suffix (ex. _a) to
the bootloader, and this is what this version of cboot expects.
However, new versions drop the underscore, wreaking havoc on the
slot setting code. Detect if the suffix passed in is the legacy
type or if its the new type, and if it is, then just prepend an
underscore so we can keep using the same existing SMD code.